python/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

